requirements.txt
pandas 
numpy 
requests
seaborn
matplotlib
scikit-learn
plotly
joblib
treamlit